requirements.txt
python-dotenv
langchain
langchain_openai
langchain-classic
# chromadb
langchain_community
pypdf2